pyproject.toml
[project]
name = "choose-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.6"
dependencies = [
    "fastmcp>=0.4.1",
    "google-cloud-bigquery-storage>=2.28.0",
    "langchain-community>=0.3.18",
    "sqlalchemy-bigquery>=1.12.1",
]
